refactor: standardize drawable icons to 24dp, fix layout tools:context paths
