+++ app/src/main/java/com/drs/auralife/presentation/explore/adapter/ExploreFilmAdapter.kt
@@ -5,8 +5,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.drs.auralife.R
 import com.drs.auralife.presentation.common.MyAppGlideModule
@@ -14,7 +14,12 @@ import com.drs.auralife.domain.model.Film
 
 class ExploreFilmAdapter(
     private val onItemClick: (String) -> Unit,
-) : ListAdapter<Film, ExploreFilmAdapter.ViewHolder>(DiffCallback) {
+) : RecyclerView.Adapter<ExploreFilmAdapter.ViewHolder>() {
+
+    private val differ = AsyncListDiffer(this, object : DiffUtil.ItemCallback<Film>() {
+        override fun areItemsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem.slug == newItem.slug
+        override fun areContentsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem == newItem
+    })
 
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val ivPoster: ImageView = itemView.findViewById(R.id.posterView)
@@ -28,19 +33,28 @@ class ExploreFilmAdapter(
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val film = getItem(position)
+        val film = differ.currentList[position]
         holder.apply {
             tvTitle.textAlignment = View.TEXT_ALIGNMENT_CENTER
             tvDetails.textAlignment = View.TEXT_ALIGNMENT_CENTER
             tvTitle.text = film.title
-            tvDetails.text = film.category
             MyAppGlideModule.loadImage(itemView.context, film.posterUrl, ivPoster)
             itemView.setOnClickListener { onItemClick(film.slug) }
         }
     }
 
-    private object DiffCallback : DiffUtil.ItemCallback<Film>() {
-        override fun areItemsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem.slug == newItem.slug
-        override fun areContentsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem == newItem
+    override fun getItemCount(): Int = differ.currentList.size
+
+    fun addItem(newItems: List<Film>) {
+        val current = differ.currentList.toMutableList()
+        newItems.forEach { film ->
+            val existingIndex = current.indexOfFirst { it.slug == film.slug }
+            if (existingIndex != -1) {
+                current[existingIndex] = film
+            } else {
+                current.add(film)
+            }
+        }
+        differ.submitList(current)
     }
 }
