+++ app/src/main/java/com/drs/auralife/presentation/explore_details/adapter/ExploreDetailFilmAdapter.kt
@@ -1,25 +1,20 @@
-package com.drs.auralife.presentation.explore.adapter
+package com.drs.auralife.presentation.explore_details.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.drs.auralife.R
-import com.drs.auralife.presentation.common.MyAppGlideModule
 import com.drs.auralife.domain.model.Film
+import com.drs.auralife.presentation.common.MyAppGlideModule
 
-class CategoryFilmAdapter(
+class ExploreDetailFilmAdapter(
     private val onItemClick: (String) -> Unit,
-) : RecyclerView.Adapter<CategoryFilmAdapter.ViewHolder>() {
-
-    private val differ = AsyncListDiffer(this, object : DiffUtil.ItemCallback<Film>() {
-        override fun areItemsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem.slug == newItem.slug
-        override fun areContentsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem == newItem
-    })
+) : ListAdapter<Film, ExploreDetailFilmAdapter.ViewHolder>(DiffCallback) {
 
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val ivPoster: ImageView = itemView.findViewById(R.id.posterView)
@@ -33,29 +28,18 @@ class CategoryFilmAdapter(
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val film = differ.currentList[position]
+        val film = getItem(position)
         holder.apply {
             tvTitle.textAlignment = View.TEXT_ALIGNMENT_CENTER
             tvDetails.textAlignment = View.TEXT_ALIGNMENT_CENTER
             tvTitle.text = film.title
-            tvDetails.text = film.category
             MyAppGlideModule.loadImage(itemView.context, film.posterUrl, ivPoster)
             itemView.setOnClickListener { onItemClick(film.slug) }
         }
     }
 
-    override fun getItemCount(): Int = differ.currentList.size
-
-    fun addItem(newItems: List<Film>) {
-        val current = differ.currentList.toMutableList()
-        newItems.forEach { film ->
-            val existingIndex = current.indexOfFirst { it.slug == film.slug }
-            if (existingIndex != -1) {
-                current[existingIndex] = film
-            } else {
-                current.add(film)
-            }
-        }
-        differ.submitList(current)
+    private object DiffCallback : DiffUtil.ItemCallback<Film>() {
+        override fun areItemsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem.slug == newItem.slug
+        override fun areContentsTheSame(oldItem: Film, newItem: Film): Boolean = oldItem == newItem
     }
 }
