+++ app/src/main/res/drawable/ic_add_library_24.xml
