+++ app/src/main/res/drawable/ic_facebook_24.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <group>
+        <clip-path android:pathData="M0,0h24v24h-24z" />
+        <path
+            android:fillColor="#3C5A9A"
+            android:pathData="M12,24C18.628,24 24,18.628 24,12C24,5.372 18.628,0 12,0C5.372,0 0,5.372 0,12C0,18.628 5.372,24 12,24Z" />
+        <path
+            android:fillColor="#ffffff"
+            android:pathData="M15.901,3.684H13.243C11.666,3.684 10.311,4.347 10.311,6.634C10.319,7.43 10.311,8.194 10.311,9.052H8.086V11.957H9.968V20.316H13.425V11.901H15.707L15.914,9.044H13.365C13.365,9.044 13.371,7.773 13.365,7.404C13.365,6.5 14.306,6.552 14.362,6.552C14.81,6.552 15.68,6.554 15.903,6.552V3.684H15.901Z" />
+    </group>
+</vector>
