+++ app/src/main/res/drawable/ic_google_24.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="25dp"
-    android:height="25dp"
+    android:width="24dp"
+    android:height="24dp"
     android:viewportWidth="25"
     android:viewportHeight="25">
     <group>
