+++ app/src/main/res/drawable/ic_play_24.xml
@@ -1,5 +1,5 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
+    android:width="24dp"
     android:height="24dp"
     android:viewportWidth="20"
     android:viewportHeight="24">
