--- app/src/main/res/drawable/ic_play_pause.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/ic_pause" android:state_selected="true" />
-    <item android:drawable="@drawable/ic_play" />
-</selector>
