+++ app/src/main/res/drawable/ic_play_pause_24.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_pause_24" android:state_selected="true" />
+    <item android:drawable="@drawable/ic_play_24" />
+</selector>
