+++ app/src/main/res/drawable/ic_rotate_24.xml
