+++ app/src/main/res/layout/app_bar.xml
@@ -31,7 +31,7 @@
         android:layout_margin="7dp"
         android:background="@color/transparent"
         android:contentDescription="@string/profile"
-        android:src="@drawable/ic_profile"
+        android:src="@drawable/ic_profile_24"
         android:scaleType="fitCenter"
         app:round="250dp"
         tools:ignore="TouchTargetSizeCheck" />
@@ -41,7 +41,7 @@
         android:layout_width="50dp"
         android:layout_height="50dp"
         android:padding="15dp"
-        android:src="@drawable/ic_search"
+        android:src="@drawable/ic_search_24"
         android:background="@color/transparent"
         android:contentDescription="@string/search"
         android:layout_gravity="end"/>
@@ -54,6 +54,6 @@
         android:padding="15dp"
         android:background="@color/transparent"
         android:contentDescription="@string/notifications"
-        android:src="@drawable/ic_notifications"
+        android:src="@drawable/ic_notifications_24"
         android:visibility="gone"/>
 </androidx.appcompat.widget.Toolbar>
