+++ app/src/main/res/layout/exoplayer_controller.xml
@@ -23,7 +23,7 @@
             android:contentDescription="@string/previous_video"
             android:padding="17dp"
             android:rotation="180"
-            android:src="@drawable/ic_next" />
+            android:src="@drawable/ic_next_24" />
 
         <ImageButton
             android:id="@+id/btn_rewind"
@@ -35,7 +35,7 @@
             android:contentDescription="@string/rewind"
             android:padding="17dp"
             android:rotation="180"
-            android:src="@drawable/ic_forward" />
+            android:src="@drawable/ic_forward_24" />
 
         <ImageButton
             android:id="@+id/btn_play_pause"
@@ -46,7 +46,7 @@
             android:background="@drawable/rounded_shadow"
             android:contentDescription="@string/play_pause"
             android:padding="14dp"
-            android:src="@drawable/ic_play_pause" />
+            android:src="@drawable/ic_play_pause_24" />
 
         <ImageButton
             android:id="@+id/btn_forward"
@@ -57,7 +57,7 @@
             android:background="@drawable/rounded_shadow"
             android:contentDescription="@string/forward"
             android:padding="17dp"
-            android:src="@drawable/ic_forward" />
+            android:src="@drawable/ic_forward_24" />
 
         <ImageButton
             android:id="@+id/btn_next"
@@ -68,7 +68,7 @@
             android:background="@drawable/rounded_shadow"
             android:contentDescription="@string/next_video"
             android:padding="17dp"
-            android:src="@drawable/ic_next" />
+            android:src="@drawable/ic_next_24" />
     </LinearLayout>
 
     <androidx.media3.ui.DefaultTimeBar
@@ -123,7 +123,7 @@
                 android:layout_height="wrap_content"
                 android:background="@color/transparent"
                 android:padding="12dp"
-                android:src="@drawable/ic_rotate"
+                android:src="@drawable/ic_rotate_24"
                 tools:ignore="ContentDescription, SpeakableTextPresentCheck" />
 
             <ImageButton
