+++ app/src/main/res/layout/fragment_explore.xml
@@ -6,23 +6,28 @@
     android:contentDescription="@string/explore"
     tools:context=".presentation.explore.ExploreFragment">
 
-    <ScrollView
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:orientation="vertical">
 
-        <LinearLayout
-            android:id="@+id/exploreFragmentBody"
+        <FrameLayout
+            android:id="@+id/appBar"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
+            android:layout_height="50dp" />
 
-            <FrameLayout
-                android:id="@+id/appBar"
+        <ScrollView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <LinearLayout
+                android:id="@+id/exploreFragmentBody"
                 android:layout_width="match_parent"
-                android:layout_height="50dp" />
+                android:layout_height="wrap_content"
+                android:orientation="vertical" />
 
-        </LinearLayout>
-    </ScrollView>
+        </ScrollView>
+    </LinearLayout>
 
     <ProgressBar
         android:id="@+id/loading_indicator"
