+++ app/src/main/res/layout/fragment_explore_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.explore_details.ExploreDetailsFragment">
+    tools:context=".presentation.explore_details.ExploreDetailFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
@@ -12,21 +12,28 @@
         android:background="?attr/colorPrimary"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light">
 
-        <ImageView
-            android:layout_width="50dp"
+        <ImageButton
+            android:id="@+id/back_button"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/bg_logo"
-            tools:ignore="ContentDescription" />
+            android:paddingStart="0dp"
+            android:paddingTop="15dp"
+            android:paddingBottom="15dp"
+            android:paddingEnd="15dp"
+            android:src="@drawable/ic_arrow_back_24"
+            android:background="@color/transparent"
+            android:contentDescription="@string/back"/>
 
         <TextView
-            android:id="@+id/tv_name_app"
+            android:id="@+id/tv_category_name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:fontFamily="cursive"
-            android:text="@string/app_name"
+            android:text="@string/category"
             android:textColor="@color/red"
             android:textSize="20sp"
-            android:textStyle="bold"/>
+            android:textStyle="bold"
+            android:layout_gravity="center" />
     </androidx.appcompat.widget.Toolbar>
 
     <androidx.recyclerview.widget.RecyclerView
