+++ app/src/main/res/layout/fragment_home.xml
@@ -64,6 +64,6 @@
         android:layout_margin="16dp"
         android:backgroundTint="@color/shadow_color"
         android:contentDescription="@string/scroll_to_top"
-        android:src="@drawable/ic_arrow_up"
+        android:src="@drawable/ic_arrow_up_24"
         android:visibility="gone" />
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
