+++ app/src/main/res/layout/fragment_login.xml
@@ -50,7 +50,7 @@
                         android:layout_marginTop="40dp"
                         android:autofillHints="emailAddress"
                         android:background="@drawable/rounded"
-                        android:drawableStart="@drawable/ic_email"
+                        android:drawableStart="@drawable/ic_email_24"
                         android:drawablePadding="10dp"
                         android:elevation="5dp"
                         android:ems="10"
@@ -69,7 +69,7 @@
                         android:layout_marginVertical="16dp"
                         android:autofillHints="password"
                         android:background="@drawable/rounded"
-                        android:drawableStart="@drawable/ic_password_1"
+                        android:drawableStart="@drawable/ic_password_24"
                         android:drawablePadding="10dp"
                         android:elevation="5dp"
                         android:ems="10"
