+++ app/src/main/res/layout/fragment_register.xml
@@ -40,7 +40,7 @@
                     android:layout_marginTop="40dp"
                     android:autofillHints="emailAddress"
                     android:background="@drawable/rounded"
-                    android:drawableStart="@drawable/ic_email"
+                    android:drawableStart="@drawable/ic_email_24"
                     android:drawablePadding="10dp"
                     android:elevation="5dp"
                     android:ems="10"
@@ -59,7 +59,7 @@
                     android:layout_marginTop="16dp"
                     android:autofillHints="password"
                     android:background="@drawable/rounded"
-                    android:drawableStart="@drawable/ic_password_1"
+                    android:drawableStart="@drawable/ic_password_24"
                     android:drawablePadding="10dp"
                     android:elevation="5dp"
                     android:ems="10"
@@ -78,7 +78,7 @@
                     android:layout_marginVertical="16dp"
                     android:autofillHints="password"
                     android:background="@drawable/rounded"
-                    android:drawableStart="@drawable/ic_password_1"
+                    android:drawableStart="@drawable/ic_password_24"
                     android:drawablePadding="10dp"
                     android:elevation="5dp"
                     android:ems="10"
