+++ app/src/main/res/layout/item_notification.xml
@@ -16,7 +16,7 @@
         android:id="@+id/tvIconButton"
         android:layout_width="48dp"
         android:layout_height="48dp"
-        android:src="@drawable/ic_delete"
+        android:src="@drawable/ic_delete_24"
         android:background="@color/transparent"
         android:contentDescription="@string/delete"
         android:layout_gravity="end"/>
