+++ app/src/main/res/layout/nav_header.xml
@@ -12,7 +12,7 @@
         android:layout_width="75dp"
         android:layout_height="75dp"
         android:contentDescription="@string/profile"
-        android:src="@drawable/ic_profile"
+        android:src="@drawable/ic_profile_24"
         android:layout_marginTop="20dp"
         app:round="250dp"
         app:layout_constraintEnd_toEndOf="parent"
