+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,18 +2,18 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
         android:id="@+id/home"
-        android:icon="@drawable/ic_home"
+        android:icon="@drawable/ic_home_24"
         android:title="@string/home" />
     <item
         android:id="@+id/explore"
-        android:icon="@drawable/ic_explore"
+        android:icon="@drawable/ic_explore_24"
         android:title="@string/explore" />
     <item
         android:id="@+id/library"
-        android:icon="@drawable/ic_library"
+        android:icon="@drawable/ic_library_24"
         android:title="@string/library" />
     <item
         android:id="@+id/history"
-        android:icon="@drawable/ic_history"
+        android:icon="@drawable/ic_history_24"
         android:title="@string/history" />
 </menu>
